pyproject.toml
[tool.poetry]
name = "selfmailbot"
version = "2024.01.10"
description = ""
authors = ["Fedor Borshev <fedor@borshev.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
python-telegram-bot = "^20.7"
celery = "^5.0"
flask = "^3"
sentry-sdk = "~1"
pystmark = "~0.5.3"
peewee = "^3.17.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-flask = "^1.3.0"
pytest-env = "^1.1.3"
ipython = "^8.19.0"
mypy = "^1.8.0"


[tool.ruff]
line-length = 160

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "A002",
  "A003",
  "ARG",
  "ANN",
  "COM812",
  "D",
  "D211",
  "D213",
  "EM101",
  "FA102",
  "ISC001",
  "N818",
  "PLR2004",
  "S101",
  "S701",
  "PT",
]
fixable = ["ALL"]

[tool.ruff.per-file-ignores]
"tests/*.py" = [
  "ANN",
  "D",
  "S",
  "PLR",
  "RET",
]
